% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.reml.meta.R
\name{delta.reml.meta}
\alias{delta.reml.meta}
\title{Function to compute and test random-effects meta-analysis summary of delta value estimates for a single marker}
\usage{
delta.reml.meta(
  delta = NULL,
  sd.delta = NULL,
  epsilon = NULL,
  alpha = 0.05,
  alternative = "two.sided",
  tol = 1e-10,
  verbose = FALSE
)
}
\arguments{
\item{delta}{numeric vector of delta values per study}

\item{sd.delta}{numeric vector of standard error of delta values per study}

\item{epsilon}{numeric non-inferiority margin for testing cross-study validity}

\item{alpha}{numeric significance level of test}

\item{alternative}{character giving the alternative hypothesis type for testing the summary effect.
One of \code{c("less","two.sided")}, where "less" corresponds to a non-inferiority test and "two.sided"
corresponds to a two one-sided test procedure. Default is "two.sided".}

\item{tol}{numeric convergence tolerance for finding a root of the score equation}

\item{verbose}{logical flag indicating whether messages should be printed, defaults to \code{FALSE}}
}
\value{
a list with elements \itemize{
\item \code{n.studies} : numeric, number of studies considered
\item \code{tau2} : numeric, estimated tau-squared (between-study heterogeneity)
\item \code{mu.delta} : numeric, estimated mean of distribution of delta
\item \code{se.delta} : numeric, standard error of delta summary estimate with Hartung-Knapp adjustment
\item \code{ci.delta.upper} : numeric, upper (1-alpha)*100\% confidence interval for mean of delta
\item \code{ci.delta.lower} : numeric, lower (1-alpha)*100\% confidence interval for mean of delta
\item \code{p.lower} : numeric, if \code{alternative} is \code{"two.sided"}, gives the p-value corresponding to
testing the null hypothesis that \code{delta} is greater than \code{-epsilon}.
Value is \code{NULL} if \code{alternative} is \code{"less"}.
\item \code{p.upper} : numeric, if \code{alternative} is \code{"two.sided"}, gives the p-value corresponding to
testing the null hypothesis that \code{delta} is less than \code{epsilon}.
Value is \code{NULL} if \code{alternative} is \code{"less"}.
\item \code{p} : numeric, consensus p-value for hypothesis test for either the two-one-sided test or
the non-inferiorty test.
\item \code{Q} : numeric, Cochran's Q-statistic for heterogeneity between studies
\item \code{I2} : numeric, Higgins-Thompson I-squared statistic representing the total percentage of variation
attributable to between-study heterogeneity
\item \code{weights.tau} : numeric vector of raw study weights for the summary measure
\item \code{weights.tau.relative} : numeric vector of relative study weights for the summary measure,
such that each weight is a percentage adding to 100\%
\item \code{weights.tau.sum} : numeric, sum of \code{weights.tau}
}
}
\description{
Function to compute and test random-effects meta-analysis summary of delta value estimates for a single marker
}
\author{
Arthur Hughes
}
