% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.highdim.multistudy.ipd.R
\docType{data}
\name{example.data.highdim.multistudy.ipd}
\alias{example.data.highdim.multistudy.ipd}
\title{High‑dimensional multi-study individual participant surrogate candidate example dataset}
\format{
A list containing :
\describe{
\item{y1}{primary response in treated}
\item{y0}{primary response in untreated}
\item{s1}{1000 surrogate candidates in treated}
\item{s0}{1000 surrogate candidates in untreated}
\item{study1}{study names for treated}
\item{study0}{study names for untreated}
\item{hyp}{for each surrogate, \code{null false} if the surrogate is valid}
}
}
\source{
Simulated for package examples.
}
\usage{
data("example.data.highdim.multistudy.ipd", package = "SurrogateRank")
}
\description{
A simulated high‑dimensional dataset for demonstrating the RISE-Meta methodology
implemented in \pkg{SurrogateRank}. The data contains primary response and
100 surrogate candidates from 25 treated individuals and 25 untreated individuals across
5 different studies,
where 10\% of the surrogate candidates are "valid".
}
\examples{
data("example.data.highdim.multistudy.ipd", package = "SurrogateRank")
head(example.data.highdim.multistudy.ipd)

}
\keyword{datasets}
