% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.highdim.multistudy.R
\docType{data}
\name{example.data.highdim.multistudy}
\alias{example.data.highdim.multistudy}
\title{High‑dimensional, multi-study surrogate candidate example dataset}
\format{
A list containing :
\describe{
\item{y1}{primary response in treated}
\item{y0}{primary response in untreated}
\item{s1}{\code{P} surrogate candidates in treated}
\item{s0}{\code{P} surrogate candidates in untreated}
\item{study}{vector of \code{K} study names}
\item{hyp}{vector of \code{P} hypotheses where \code{null false} if the surrogate is valid}
}
}
\source{
Simulated for package examples.
}
\usage{
data("example.data.highdim.multistudy", package = "SurrogateRank")
}
\description{
A simulated high‑dimensional, multi-study dataset for demonstrating the RISE-meta methodology
implemented in \pkg{SurrogateRank}. The data contains primary response and
1000 surrogate candidates from 10 studues of varying sample sizes.
10\% of the surrogate candidates are "valid", where the strength of the surrogacy depends on the study.
}
\examples{
data("example.data.highdim.multistudy", package = "SurrogateRank")
head(example.data.highdim.multistudy)

}
\keyword{datasets}
