% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.example.data.highdim.multistudy.ipd.R
\name{generate.example.data.highdim.multistudy.ipd}
\alias{generate.example.data.highdim.multistudy.ipd}
\title{Generate multi-study individual participant data for high-dimensional surrogate candidates and response}
\usage{
generate.example.data.highdim.multistudy.ipd(
  M,
  n1,
  n0,
  p,
  prop_valid,
  valid_sigma = 1,
  corr = 0,
  mode = "simple",
  y0_mean = 0,
  y0_sd = 1,
  y1_mean = 3,
  y1_sd = 1,
  s0_mean = 0,
  s0_sd = 1,
  s1_mean = 0,
  s1_sd = 1,
  seed = 12345
)
}
\arguments{
\item{M}{number of studies}

\item{n1}{positive numeric giving the sample size in the treated groups}

\item{n0}{positive numeric giving the sample size in the untreated groups}

\item{p}{positive numeric giving the number of markers to generate}

\item{prop_valid}{numeric between 0 and 1 (inclusive) giving the proportion of surrogate
candidates to generate as valid.}

\item{valid_sigma}{non-negative numeric giving the standard deviation for valid candidates}

\item{corr}{non-negative numeric giving the correlation between the surrogate candidates}

\item{mode}{character taking values in c("simple", "complex"). If "simple", generates all variables with
(multivariate) normal distributions. Else, uses a more complex exponential distribution.}

\item{y0_mean}{numeric giving the mean of the primary endpoint in the untreated group}

\item{y0_sd}{non-negative numeric giving the standard deviation of the primary endpoint in the untreated group}

\item{y1_mean}{numeric giving the mean of the primary endpoint in the treated group}

\item{y1_sd}{non-negative numeric giving the standard deviation of the primary endpoint in the treated group}

\item{s0_mean}{numeric giving the mean of the surrogate candidates in the untreated group}

\item{s0_sd}{non-negative numeric giving the standard deviation of the surrogate candidates in the untreated group}

\item{s1_mean}{numeric giving the mean of the surrogate candidates in the treated group}

\item{s1_sd}{non-negative numeric giving the standard deviation of the surrogate candidates in the treated group}

\item{seed}{numeric giving a seed for reproducibility}
}
\value{
A list with the following components:
\describe{
\item{y1}{vector containing primary endpoint values in treated group}
\item{y0}{vector containing primary endpoint values in untreated group}
\item{s1}{n1 times p matrix containing surrogate candidate values in treated group}
\item{s0}{n0 times p matrix containing surrogate candidate values in untreated group}
\item{study1}{study names for treated samples}
\item{study0}{study names for untreated samples}
\item{hyp}{character vector giving the truth behind the null hypothesis for each surrogate candidate}
}
}
\description{
Generates individual participant data for high-dimensional surrogate candidates using one
of two data generating processes, as described in Hughes A et al (2025) \url{doi:10.1002/sim.70241}.
}
\examples{
res <- generate.example.data.highdim.multistudy.ipd(M = 5, n1 = 25, n0 = 25, p = 500, prop_valid = 1)
dim(res$s1)       # (5 studies x 25 individuals = 125) x 500
}
