% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rise.evaluate.meta.R
\name{rise.evaluate.meta}
\alias{rise.evaluate.meta}
\title{Function to perform the evaluation stage of RISE-meta : Meta-Analysis of High-Dimensional Surrogate Markers}
\usage{
rise.evaluate.meta(
  yone,
  yzero,
  sone,
  szero,
  studyone,
  studyzero,
  alpha = 0.05,
  power.want.s.study = NULL,
  epsilon.study = NULL,
  epsilon.meta.mode = "user",
  epsilon.meta = NULL,
  u.y.hyp = NULL,
  p.correction = "BH",
  n.cores = 1,
  alternative = "two.sided",
  test = "knha",
  paired.all = FALSE,
  paired.studies = NULL,
  evaluate.weights = TRUE,
  screening.weights = NULL,
  weight.mode = "diff.epsilon",
  markers = NULL,
  return.all.evaluate = FALSE,
  return.forest.plot = TRUE,
  return.fit.plot = TRUE,
  show.pooled.effect = TRUE,
  meta.analysis.method = "RE"
)
}
\arguments{
\item{yone}{numeric vector of primary response values in the treated participants}

\item{yzero}{numeric vector of primary response values in the untreated participants}

\item{sone}{matrix or dataframe of surrogate candidates in the treated group with dimension
\code{n1 x p} where n1 is the number of treated samples and p the number of candidates. Sample
ordering must match exactly yone.}

\item{szero}{matrix or dataframe of surrogate candidates in the untreated group with dimension
\code{n0 x p} where n0 is the number of treated samples and p the number of candidates. Sample
ordering must match exactly yzero.}

\item{studyone}{character vector of length \code{n1} indicating the study corresponding to each treated sample.
Ordering much match yone.}

\item{studyzero}{character vector of length \code{n0} indicating the study corresponding to each untreated sample.
Ordering much match yzero.}

\item{alpha}{significance level for determining valid surrogates. Default is \code{0.05}.}

\item{power.want.s.study}{numeric in (0,1) - power desired for a test of treatment effect based on the
surrogate candidate. If \code{return.all.evaluate = TRUE}, either this or \code{epsilon.study} argument must be specified.}

\item{epsilon.study}{numeric in (0,1) - non-inferiority margin for determining surrogate validity in the
within-study screening phase. If \code{return.all.evaluate = TRUE}, either this or \code{power.want.s.study} argument must be specified.}

\item{epsilon.meta.mode}{character string specifying the mode to choose the value of the acceptable margin defined
by epsilon. By default, this is set to "user", where the value of epsilon is fixed by the user, defined by the
value of the argument \code{epsilon.meta}. The alternative is to set this as "mean.power", which corresponds to
taking the mean value of epsilon across studies such that the power to detect departures from the null within
each study is defined by the \code{power.want.s.study} argument.}

\item{epsilon.meta}{numeric in (0,1) - non-inferiority margin for determining surrogate validity
in the meta-analysis stage. Must be specified.}

\item{u.y.hyp}{hypothesised value of the treatment effect on the primary response on the probability
scale. If not given, it will be estimated based on the observations.}

\item{p.correction}{character. Method for p-value adjustment (see \code{p.adjust()} function).
Defaults to the Benjamini-Hochberg method (\code{"BH"}).}

\item{n.cores}{numeric giving the number of cores to commit to parallel computation in order to
improve computational time through the \code{pbmcapply()} function. Defaults to \code{1}.}

\item{alternative}{character giving the alternative hypothesis type. One of
\code{c("less","two.sided")}, where "less" corresponds to a non-inferiority test and "two.sided"
corresponds to a two one-sided test procedure. Default is "two.sided".}

\item{test}{character giving the type of test to be performed. The default is \code{knha}, corresponding to
variance estimation using the more conservative Hartung-Knapp estimator and performes tests with the t-distribution,
whereas setting this argument to \code{z} estimates the variance with the conventional estimator and uses a normal approximation for testing.}

\item{paired.all}{logical flag giving if the data is independent or paired. If \code{FALSE} (default),
samples are assumed independent. If \code{TRUE}, all samples are assumed to be from a paired design.
The pairs are specified by matching the rows of \code{yone} and \code{sone} to the rows of
\code{yzero} and \code{szero}, and all studies must be paired. If only some studies are paired and
others have independent samples, one may specify the \code{paired.studies} argument instead.}

\item{paired.studies}{character vector specifying the names of the studies in \code{studyone} or \code{studyzero}
which are paired, in the case where some have paired designs and others do not. By default, this is \code{NULL},
indicating that study designs are all specified by the \code{paired.all} argument.}

\item{evaluate.weights}{logical flag. If \code{TRUE} (default), the composite surrogate is
constructed with weights such that surrogates which are
predicted to be stronger receive more weight.}

\item{screening.weights}{dataframe with columns \code{marker} and \code{weight} giving the weight
in for the evaluation. Typically this is taken directly from the
screening stage as the output from the \code{rise.screen.meta()} function.
Must be given if \code{evaluate.weights} is \code{TRUE}.}

\item{weight.mode}{character giving the type of weighting to return to be used in case \code{return.all.evaluate = TRUE}.
See \code{rise.screen.meta()} for detail.}

\item{markers}{a vector of marker names (column names of szero and sone) to evaluate.
If not given, will default to evaluating all markers in the dataframes.}

\item{return.all.evaluate}{logical flag. If \code{TRUE}, a dataframe will be returned
giving the meta-analysis evaluation of each individual marker passed to the
evaluation stage. Defaults to \code{FALSE} for computational time.}

\item{return.forest.plot}{logical flag. If \code{TRUE} (default), a forest plot of the effect sizes for the
combined signature across studies, with its meta-analysis summary measure, will be included in the output.}

\item{return.fit.plot}{logical flag. If \code{TRUE} (default), a plot of the effects on the primary response
versus the effects on the combined surrogate signature for each study will be included in the output.}

\item{show.pooled.effect}{logical flag. If \code{TRUE} (default), the forest plot will show the pooled effect
estimate. Otherwise, it will just show the individual trial estimates.}

\item{meta.analysis.method}{character giving the meta-analysis method to be used. The default is \code{RE}, corresponding to
random-effects meta-analysis, whereas setting this argument to \code{FE} uses fixed-effects meta-analysis.}
}
\value{
a list with elements \itemize{
\item \code{individual.metrics} : if \code{return.all.evaluate}=\code{TRUE}, a list containing
dataframes \code{individual.metrics.study} (per-study results for individual markers) and
\code{individual.metrics.meta} (meta-analysis results for individual markers).
\item \code{evaluation.metrics.study} : study-level results for the combined marker, gamma.
\item \code{evaluation.metrics.meta} : meta-analysis results for the combined marker, gamma.
\item \code{gamma.s} : a list with elements \code{gamma.s.one} and \code{gamma.s.zero}, giving
the values of the combined surrogate marker gamma in the treated and untreated groups, respectively.
\item \code{gamma.s.plot} : if \code{return.forest.plot} and/or \code{return.fit.plot} are \code{TRUE},
returns evaluation plots as a list
}
}
\description{
Function to perform the evaluation stage of RISE-meta : Meta-Analysis of High-Dimensional Surrogate Markers
}
\examples{
data("example.data.highdim.multistudy.ipd")
yone <- example.data.highdim.multistudy.ipd$y1
yzero <- example.data.highdim.multistudy.ipd$y0
sone <- example.data.highdim.multistudy.ipd$s1
szero <- example.data.highdim.multistudy.ipd$s0
studyone <- example.data.highdim.multistudy.ipd$study1
studyzero <- example.data.highdim.multistudy.ipd$study0
rise.meta.screen.result <- rise.screen.meta(yone, yzero, sone, szero, studyone, studyzero, epsilon.study = 0.2, epsilon.meta = 0.2, n.cores = 12)
markers = rise.meta.screen.result[["significant.markers"]]
screening.weights = rise.meta.screen.result[["screening.weights"]]
rise.meta.evaluate.result <- rise.evaluate.meta(yone, yzero, sone, szero, studyone, studyzero, epsilon.meta = 0.2, markers = markers, screening.weights = screening.weights, epsilon.study = 0.2, n.cores = 12)
}
\author{
Arthur Hughes
}
