% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rise.screen.meta.R
\name{rise.screen.meta}
\alias{rise.screen.meta}
\title{Function to perform the screening stage of RISE-meta : Meta-Analysis of High-Dimensional Surrogate Markers}
\usage{
rise.screen.meta(
  yone,
  yzero,
  sone,
  szero,
  studyone,
  studyzero,
  alpha = 0.05,
  power.want.s.study = NULL,
  epsilon.study = NULL,
  epsilon.meta.mode = "user",
  epsilon.meta = NULL,
  u.y.hyp = NULL,
  p.correction = "BH",
  n.cores = 1,
  alternative = "two.sided",
  test = "knha",
  paired.all = FALSE,
  paired.studies = NULL,
  return.all.screen = TRUE,
  return.all.weights = FALSE,
  weight.mode = "diff.epsilon",
  return.screen.plot = TRUE,
  screen.plot.topN = 15,
  screen.plot.point.estimate = FALSE,
  normalise.weights = TRUE,
  return.forest.plot = TRUE,
  return.fit.plot = TRUE,
  show.pooled.effect = TRUE,
  return.study.similarity.plot = TRUE,
  return.evaluate.results = TRUE,
  meta.analysis.method = "RE"
)
}
\arguments{
\item{yone}{numeric vector of primary response values in the treated participants}

\item{yzero}{numeric vector of primary response values in the untreated participants}

\item{sone}{matrix or dataframe of surrogate candidates in the treated group with dimension
\code{n1 x p} where n1 is the number of treated samples and p the number of candidates. Sample
ordering must match exactly yone.}

\item{szero}{matrix or dataframe of surrogate candidates in the untreated group with dimension
\code{n0 x p} where n0 is the number of treated samples and p the number of candidates. Sample
ordering must match exactly yzero.}

\item{studyone}{character vector of length \code{n1} indicating the study corresponding to each treated sample.
Ordering much match yone.}

\item{studyzero}{character vector of length \code{n0} indicating the study corresponding to each untreated sample.
Ordering much match yzero.}

\item{alpha}{significance level for determining surrogate candidates in both stages. Default is \code{0.05}.}

\item{power.want.s.study}{numeric in (0,1) - power desired for a test of treatment effect based on the
surrogate candidate. Either this or \code{epsilon.study} argument must be specified.}

\item{epsilon.study}{numeric in (0,1) - non-inferiority margin for determining surrogate validity in the
within-study screening phase. Either this or \code{power.want.s.study} argument must be specified.}

\item{epsilon.meta.mode}{character string specifying the mode to choose the value of the acceptable margin defined
by epsilon. By default, this is set to "user", where the value of epsilon is fixed by the user, defined by the
value of the argument \code{epsilon.meta}. The alternative is to set this as "mean.power", which corresponds to
taking the mean value of epsilon across studies such that the power to detect departures from the null within
each study is defined by the \code{power.want.s.study} argument.}

\item{epsilon.meta}{numeric in (0,1) - fixed non-inferiority margin for determining surrogate validity
in the meta-analysis stage.}

\item{u.y.hyp}{hypothesised value of the treatment effect on the primary response on the probability
scale. If not given, it will be estimated based on the observations.}

\item{p.correction}{character. Method for p-value adjustment (see \code{p.adjust()} function).
Defaults to the Benjamini-Hochberg method (\code{"BH"}).}

\item{n.cores}{numeric giving the number of cores to commit to parallel computation in order to
improve computational time through the \code{pbmcapply()} function. Defaults to \code{1}.}

\item{alternative}{character giving the alternative hypothesis type. One of
\code{c("less","two.sided")}, where "less" corresponds to a non-inferiority test and "two.sided"
corresponds to a two one-sided test procedure. Default is "two.sided".}

\item{test}{character giving the type of test to be performed. The default is \code{knha}, corresponding to
variance estimation using the more conservative Hartung-Knapp estimator and performes tests with the t-distribution,
whereas setting this argument to \code{z} estimates the variance with the conventional estimator and uses a normal approximation for testing.}

\item{paired.all}{logical flag giving if the data is independent or paired. If \code{FALSE} (default),
samples are assumed independent. If \code{TRUE}, all samples are assumed to be from a paired design.
The pairs are specified by matching the rows of \code{yone} and \code{sone} to the rows of
\code{yzero} and \code{szero}, and all studies must be paired. If only some studies are paired and
others have independent samples, one may specify the \code{paired.studies} argument instead.}

\item{paired.studies}{character vector specifying the names of the studies in \code{studyone} or \code{studyzero}
which are paired, in the case where some have paired designs and others do not. By default, this is \code{NULL},
indicating that study designs are all specified by the \code{paired.all} argument.}

\item{return.all.screen}{logical flag. If \code{TRUE} (default), a dataframe will be returned giving
the screening results for all candidates. Else, only the significant candidates will be returned.}

\item{return.all.weights}{logical flag. If \code{FALSE} (default), a dataframe will be returned giving
weights for significant markers screened. If \code{TRUE}, weights for all markers will be returned. Note
that, if normalised weights are required, these will only be returned for significant markers, and raw
weights will be returned in a second column.}

\item{weight.mode}{character giving the type of weighting to return. One of
\code{c("diff.epsilon", "inverse.delta", or "none")}. The default is \code{"diff.epsilon"}, which calculates
weights as the proportion of the interval between 0 and epsilon.study cut by the absolute value of delta,
therefore giving delta = 0 a weight of 1 and delta = epsilon.study a weight of 0. Another option is
\code{"inverse.delta"} where the weights are determined by taking the inverse of the absolute values of delta.
When \code{"none"}, the weights are set to 1 for every marker.}

\item{return.screen.plot}{logical flag. If \code{TRUE} (default), returns a forest plot of the top predictors, sorted by p-value,
from the screening stage. The number of predictors to display is given by the \code{screen.plot.topN} argument, which has default
value 15.}

\item{screen.plot.topN}{number of predictors to display in the screening results figure, default value is 15.}

\item{screen.plot.point.estimate}{logical flag. If \code{FALSE} (default), uses the \code{screen.plot.topN} argument to determine how many
markers to display on the screen plot. Otherwise, plots all the markers with a point estimate within the equivalence region.}

\item{normalise.weights}{logical flag. If \code{TRUE} (default), the weights are normalised by the sum of
all the weights such that the maximum weight is 1, which can help with interpretability.}

\item{return.forest.plot}{logical flag. If \code{TRUE} (default), a forest plot of the effect sizes for the
combined signature across studies, with its meta-analysis summary measure and prediction interval, will be included in the output.}

\item{return.fit.plot}{logical flag. If \code{TRUE} (default), a plot of the effects on the primary response
versus the effects on the combined surrogate signature for each study will be included in the output.}

\item{show.pooled.effect}{logical flag. If \code{TRUE} (default), the forest plot will show the pooled effect
estimate. Otherwise, it will just show the individual trial estimates.}

\item{return.study.similarity.plot}{logical flag. If \code{TRUE} (default), will return two plots showing the similarity
between study-wise marker signatures (i.e., the application of RISE to each study individually, with p-value correction within-study).}

\item{return.evaluate.results}{logical flag. If \code{TRUE} (default), returns results for combined marker gamma, evaluated on the same
data. Can be useful to set this as \code{FALSE} to save computational time if this is not of interest.}

\item{meta.analysis.method}{character giving the meta-analysis method to be used. The default is \code{RE}, corresponding to
random-effects meta-analysis, whereas setting this argument to \code{FE} uses fixed-effects meta-analysis.}
}
\value{
a list with elements \itemize{
\item \code{screening.metrics.study} : dataframe of per-study results from RISE screening.
For each candidate marker - study name, study sample size, estimate of delta, standard error of delta.
\item \code{screening.metrics.meta} : dataframe of meta-analysis screening results.
For each candidate marker - number of studies \code{n.studies},
estimate of mean delta value \code{mu.delta},
its standard error \code{se.delta}, confidence interval and prediction interval,
estimate of tau-squared \code{tau2}, Cochran's Q-statistic and Higgins-Thompson I-Squared,
unadjusted and adjusted meta-analysis p-values, and standardised weights.
Note : if using the non-inferiority test (i.e. \code{alternative = "less"}),
the intervals have width (1-\code{alpha})*100\%,
whereas the two-one-sided test (i.e. \code{alternative = "two.sided"})
corresponds to a (1-2\code{alpha})*100\% width.
\item \code{significant.markers}: character vector of markers with meta-analysis p-values \code{< alpha}
\item \code{screening.weights}: dataframe giving marker names and the standardised meta-analysis weights
\item \code{evaluation.metrics.study} : dataframe of per-study results for the combined marker gamma, evaluated on the same data
\item \code{evaluation.metrics.meta} : dataframe of meta-analysis results for the combined marker gamma, evaluated on the same data
\item \code{gamma.s.plot}: if \code{return.forest.plot}, \code{return.fit.plot}, and/or  \code{return.study.similarity.plot}
are \code{TRUE}, returns fitted evaluation plots on training data as a list.
}
}
\description{
The RISE screening algorithm is applied to each study using a rank-based measure of
treatment effect similarity. In the second stage, these effect estimates are combined using a
random-effects meta-analysis and the retained markers are those for which there is strong evidence
of surrogacy across many studies.
}
\examples{
data("example.data.highdim.multistudy.ipd")
yone <- example.data.highdim.multistudy.ipd$y1
yzero <- example.data.highdim.multistudy.ipd$y0
sone <- example.data.highdim.multistudy.ipd$s1
szero <- example.data.highdim.multistudy.ipd$s0
studyone <- example.data.highdim.multistudy.ipd$study1
studyzero <- example.data.highdim.multistudy.ipd$study0
rise.meta.screen.result <- rise.screen.meta(
yone, yzero, 
sone, szero, 
studyone, studyzero, 
epsilon.study = 0.2, epsilon.meta = 0.2
)
}
\author{
Arthur Hughes
}
