% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.multi.study.surrogates.R
\name{simulate.multi.study.surrogates}
\alias{simulate.multi.study.surrogates}
\title{Title}
\usage{
\method{simulate}{multi.study.surrogates}(
  u_y_mean = 0.8,
  u_y_sd = 0.1,
  d_valid = -0.1,
  d_invalid = -0.5,
  u_s_sd_valid = 0.1,
  u_s_sd_invalid = 0.1,
  epsilon = 0.2,
  M = 10,
  J = 500,
  prop_valid = 0.1
)
}
\arguments{
\item{u_y_mean}{Mean of treatment effects on primary endpoint, bounded between 0 and 1}

\item{u_y_sd}{Standard deviation of treatment effects on primary endpoint}

\item{d_valid}{Average distance of treatment effect on valid surrogates and on primary outcome within each trial.
Must satisfy \code{|d_valid| <= epsilon}.}

\item{d_invalid}{Average distance of treatment effect on invalid surrogates and on primary outcome within each trial.
Must satisfy \code{|d_invalid| > epsilon}.}

\item{u_s_sd_valid}{Standard deviation of distance of treatment effect on valid surrogates and on primary outcome.}

\item{u_s_sd_invalid}{Standard deviation of distance of treatment effect on invalid surrogates and on primary outcome.}

\item{epsilon}{Margin for definition of valid surrogates.}

\item{M}{Number of trials.}

\item{J}{Number of surrogate candidates.}

\item{prop_valid}{Proportion of surrogate candidates which are valid.}
}
\value{
list containing elements \itemize{
\item \code{uy} : numeric vector of length \code{M} containing treatment effects on primary endpoint across trials
\item \code{us} : numeric matrix of dimension \code{M} times \code{J} containing treatment effects on each of \code{J} candidate markers.
\item \code{hyp} : vector of length \code{J} containing the truth of surrogate validity. \code{null false} corresponds to valid surrogates
whereas \code{null true} corresponds to invalid surrogates.
\item \code{epsilon} : value of epsilon used to define surrogate validity.
}
}
\description{
Title
}
\examples{
example.data.highdim.multistudy = simulate.multi.study.surrogates(u_y_mean = 0.8, u_y_sd = 0.1, d_valid = -0.1, d_invalid = -0.5, u_s_sd_valid = 0.1, u_s_sd_invalid = 0.1, epsilon = 0.2, M = 10, J = 500, prop_valid = 0.1)
}
